allow ignoreMetrics in new manager
